Sanitize request body for all HTTP verbs
The HTTP spec allows a GET body, and Rails violates the intent of the
spec by parsing that body (and I wouldn't be surprised if other
frameworks do as well). Unfortunately this means that a poorly behaved
client (looking at you, EasuoSpider) can submit bad UTF-8 in a GET and
cause upstream problems.

This fixes the issue by sanitizing `rack.input` - the request body - for
all request methods.

Closes #14.