Protect against Symbol keys in the rack `env`
Sometimes the rack `env` can get a key that's a symbol. This was causing
the `starts_with?` method to fail. This is simply fixed by added a `to_s`
call to the key before calling `starts_with?`

Test for the behavior by adding a dummy symbol key in the env