Don't use a mutable constant as Rack response
A common gotcha in Rack API is to return a constant as a response.
Another middleware higher in the stack may mutate the header hash
to add some user specific things (.e.g `Set-Cookie`) and this state
then leak across requests.

In the case of `rack-utf8-sanitizer`, the risk is limited because
it's likely to be among the very first middlewares, but it's still
best not to do this.