Fix URI::RFC2396_PARSER issue with older Rubies
Switching to URI::RFC2396_PARSER in 028fd338d8fc667c resulted in
problems with older versions of Ruby where URI::RFC2396_PARSER is not
defined. Although URI::RFC2396_PARSER was added and backported to the
vendored uri library in older versions of Ruby, we check for
URI::RFC2396_PARSER's presence such that the warnings are avoided when
using a newer non-vendored version of the gem.